refactor(back-end): Refatora lógica do Menu e reorganiza arquivos
Reestrutura a lógica de inicialização e rotas do menu, e organiza controllers/models:
1. Migração: Lógica de menu movida de 'Menu.js' e 'app.js' para o novo arquivo 'menuData.js' (src/utils).
2. Limpeza: Arquivos de Menu antigo ('menuController.js', 'menuModel.js', 'menu.js') foram movidos para novos arquivos.
3. Ajuste de rotas e controllers de usuário.
