+++ app.js
@@ -14,7 +14,7 @@ var HOST_APP = process.env.APP_HOST;
 var app = express();
 
 var indexRouter = require("./src/routes/index");
-var usuarioRouter = require("./src/routes/usuarios");
+var menuRouter = require("./src/routes/menu");
 var gerenciamentoMaquinasRouter = require("./src/routes/maquinas");
 
 app.use(express.json());
@@ -24,7 +24,7 @@ app.use(express.static(path.join(__dirname, "public")));
 app.use(cors());
 
 app.use("/", indexRouter);
-app.use("/usuarios", usuarioRouter);
+app.use("/menu", menuRouter);
 app.use("/gerenciamentoMaquinas", gerenciamentoMaquinasRouter);
