+++ src/controllers/menuController.js
@@ -1,93 +1,12 @@
-var usuarioModel = require("../models/usuarioModel");
+var menuModel = require("../models/menuModel");
 const { modulos } = require('../utils/menuData');
-function autenticar(req, res) {
-  var email = req.body.emailServer;
-  var senha = req.body.senhaServer;
-
-  if (email == undefined) {
-    res.status(400).send("Seu email está undefined!");
-  } else if (senha == undefined) {
-    res.status(400).send("Sua senha está indefinida!");
-  } else {
-    usuarioModel
-      .autenticar(email, senha)
-      .then(function (resultadoAutenticar) {
-        console.log(`\nResultados encontrados: ${resultadoAutenticar.length}`);
-        console.log(`Resultados: ${JSON.stringify(resultadoAutenticar)}`);
-
-        if (resultadoAutenticar.length == 1) {
-          console.log(resultadoAutenticar);
-
-          aquarioModel
-            .buscarAquariosPorEmpresa(resultadoAutenticar[0].empresaId)
-            .then((resultadoAquarios) => {
-              if (resultadoAquarios.length > 0) {
-                res.json({
-                  id: resultadoAutenticar[0].id,
-                  email: resultadoAutenticar[0].email,
-                  nome: resultadoAutenticar[0].nome,
-                  senha: resultadoAutenticar[0].senha,
-                  aquarios: resultadoAquarios,
-                });
-              } else {
-                res.status(204).json({ aquarios: [] });
-              }
-            });
-        } else if (resultadoAutenticar.length == 0) {
-          res.status(403).send("Email e/ou senha inválido(s)");
-        } else {
-          res.status(403).send("Mais de um usuário com o mesmo login e senha!");
-        }
-      })
-      .catch(function (erro) {
-        console.log(erro);
-        console.log(
-          "\nHouve um erro ao realizar o login! Erro: ",
-          erro.sqlMessage
-        );
-        res.status(500).json(erro.sqlMessage);
-      });
-  }
-}
-
-function cadastrar(req, res) {
-  var nome = req.body.nomeServer;
-  var email = req.body.emailServer;
-  var senha = req.body.senhaServer;
-  var fkEmpresa = req.body.idEmpresaVincularServer;
-
-  if (nome == undefined) {
-    res.status(400).send("Seu nome está undefined!");
-  } else if (email == undefined) {
-    res.status(400).send("Seu email está undefined!");
-  } else if (senha == undefined) {
-    res.status(400).send("Sua senha está undefined!");
-  } else if (fkEmpresa == undefined) {
-    res.status(400).send("Sua empresa a vincular está undefined!");
-  } else {
-    usuarioModel
-      .cadastrar(nome, email, senha, fkEmpresa)
-      .then(function (resultado) {
-        res.json(resultado);
-      })
-      .catch(function (erro) {
-        console.log(erro);
-        console.log(
-          "\nHouve um erro ao realizar o cadastro! Erro: ",
-          erro.sqlMessage
-        );
-        res.status(500).json(erro.sqlMessage);
-      });
-  }
-}
-
 function getMenu(req, res) {
   const idUsuario = req.params.idUsuario;
   if (idUsuario == undefined) {
     return res.status(400).send("Seu idUsuario está undefined!");
   }
 
-  usuarioModel.getMenu(idUsuario)
+  menuModel.getMenu(idUsuario)
     .then(function (resultado) {
       if (resultado.length === 0) {
         return res.status(404).json({ mensagem: "Usuário não encontrado." });
@@ -96,14 +15,14 @@ function getMenu(req, res) {
       const permissoesArray = permissoesDoBanco.split(';');
 
       const menu = {
-        alertaSuportePC: gerarLinkHTML(modulos.home) +
+        alertaSuportePC:
           (permissoesArray.includes('ver_alertas') ? gerarLinkHTML(modulos.alertas) : '') +
           (permissoesArray.includes('ver_suporte') ? gerarLinkHTML(modulos.suporte) : ''),
 
         painelPC: permissoesArray.includes('ver_paineis') ? gerarDropdownHTML(modulos.paineis, false) : '',
         gestaoAreaPC: gerarSecaoGestaoHTML(permissoesArray, false),
 
-        alertaSuporteMobile: gerarLinkHTML(modulos.home) +
+        alertaSuporteMobile:
           (permissoesArray.includes('ver_alertas') ? gerarLinkHTML(modulos.alertas) : '') +
           (permissoesArray.includes('ver_suporte') ? gerarLinkHTML(modulos.suporte) : ''),
 
@@ -165,7 +84,5 @@ function gerarSecaoGestaoHTML(permissoesArray) {
   return ''; 
 }
 module.exports = {
-  autenticar,
-  cadastrar,
   getMenu,
 };
