--- src/routes/usuarios.js
@@ -1,18 +1,0 @@
-var express = require("express");
-var router = express.Router();
-
-var usuarioController = require("../controllers/usuarioController");
-
-router.post("/cadastrar", function (req, res) {
-  usuarioController.cadastrar(req, res);
-});
-
-router.post("/autenticar", function (req, res) {
-  usuarioController.autenticar(req, res);
-});
-
-router.get("/getMenu/:idUsuario", function (req, res) {
-  usuarioController.getMenu(req, res);
-});
-
-module.exports = router;
