+++ src/utils/menuData.js
@@ -1,8 +1,7 @@
 const modulos = {
-    'home': { titulo: 'Home', icone: 'home_white.svg', link: 'home.html', permissao: 'ver_paineis' },
     'paineis': { titulo: 'Paineis', icone: 'painel_white.svg', link: 'painel.html', permissao: 'ver_paineis', isDropdown: true, dropdownItens: ['Painel 1', 'Painel 2', 'Painel 3'] },
     'alertas': { titulo: 'Alertas', icone: 'alertas_white.svg', link: 'alertas.html', permissao: 'ver_alertas' },
-    'suporte': { titulo: 'Suporte', icone: 'suport_white.svg', link: '#', permissao: 'ver_suporte' },
+    'suporte': { titulo: 'Suporte', icone: 'suport_white.svg', link: 'https://oberon-sis.atlassian.net/servicedesk/customer/portal/34/user/login?destination=portal%2F34', permissao: 'ver_suporte' },
     'usuarios': { titulo: 'Usuários', icone: 'users_white.svg', link: 'gerenciamentoUsuarios.html', permissao: 'gerir_usuarios' },
     'maquinas': { titulo: 'Máquinas', icone: 'maquinas_white.svg', link: 'gerenciamentoMaquinas.html', permissao: 'gerir_maquinas' },
     'empresa': { titulo: 'Empresa', icone: 'empresa_white.svg', link: 'empresa.html', permissao: 'gerir_empresa' }
